sqlAlchemy_Practice/requirements.txt
sqlalchemy
pymysql
python-dotenv

